refactor: move interface to its separate folder
